+++ examples/mini-app/src/hooks/usePlaceBet.ts
@@ -3,8 +3,13 @@ import {
   CASINO_GAME_TYPE,
   CasinoChainId,
   CoinToss,
+  casinoChainById,
   Dice,
   GenericCasinoBetParams,
+  getPlaceBetEventData,
+  getPlaceBetFunctionData,
+  getPlacedBetFromReceipt,
+  getRollEventData,
   Keno,
   KenoConfiguration,
   MAX_SELECTABLE_DICE_NUMBER,
@@ -12,14 +17,9 @@ import {
   MIN_SELECTABLE_DICE_NUMBER,
   MIN_SELECTABLE_ROULETTE_NUMBER,
   Roulette,
-  casinoChainById,
-  getPlaceBetEventData,
-  getPlaceBetFunctionData,
-  getPlacedBetFromReceipt,
-  getRollEventData,
 } from "@betswirl/sdk-core"
 import { useCallback, useEffect, useMemo, useState } from "react"
-import { Hex, decodeEventLog } from "viem"
+import { decodeEventLog, Hex } from "viem"
 import { useAccount, usePublicClient, useWaitForTransactionReceipt, useWriteContract } from "wagmi"
 import { useChain } from "../context/chainContext"
 import { useBettingConfig } from "../context/configContext"