+++ examples/node/src/functionalities/common.ts
@@ -12,35 +12,36 @@ import {
   type ChoiceInput,
   CoinToss,
   type CoinTossChoiceInput,
+  casinoChains,
+  chainById,
   Dice,
   type DiceChoiceInput,
   FORMAT_TYPE,
+  formatRawAmount,
+  formatTxnUrl,
   GAS_PRICE_TYPE,
   GAS_TOKEN_ADDRESS,
+  getBetSwirlBetUrl,
   Keno,
   type KenoChoiceInput,
+  LEADERBOARD_TYPE,
+  labelCasinoGameByType,
   type NormalCasinoPlacedBet,
   type NormalGameChoiceInput,
   Roulette,
   type RouletteChoiceInput,
   type SignedFreebet,
+  type WEIGHTED_CASINO_GAME_TYPE,
   type WeightedCasinoPlacedBet,
+  WeightedGame,
   type WeightedGameChoiceInput,
   type WeightedGameConfiguration,
-  Wheel,
-  type WheelChoiceInput,
-  casinoChains,
-  chainById,
-  formatRawAmount,
-  formatTxnUrl,
-  getBetSwirlBetUrl,
-  labelCasinoGameByType,
 } from "@betswirl/sdk-core";
-import { WagmiBetSwirlClient, initWagmiBetSwirlClient } from "@betswirl/wagmi-provider";
+import { initWagmiBetSwirlClient, WagmiBetSwirlClient } from "@betswirl/wagmi-provider";
 import { checkbox, input, select } from "@inquirer/prompts";
 import { getBalance } from "@wagmi/core";
 import chalk from "chalk";
-import { type Hash, type Hex, type TransactionReceipt, parseUnits } from "viem";
+import { type Hash, type Hex, parseUnits, type TransactionReceipt } from "viem";
 import { getWagmiConfigForAllChains, getWagmiConfigFromCasinoChain } from "../../utils";
 
 let wagmiBetSwirlClient: WagmiBetSwirlClient;
@@ -58,6 +59,10 @@ export async function _selectChain(): Promise<CasinoChain> {
     chainId: selectedChain.id,
     affiliate: process.env.AFFILIATE_ADDRESS as Hex,
     gasPriceType: GAS_PRICE_TYPE.FAST,
+    // testMode to true to use the staging API.
+    api: {
+      testMode: true,
+    },
   });
   return selectedChain;
 }
@@ -194,23 +199,25 @@ export async function _selectInput(
       });
       break;
     }
-    // Wheel
+    // Wheel & Plinko (Weighted game)
     default: {
       input = await select({
         message: "Select a configuration",
         loop: false,
         // You could bring your own configurations here by passing customConfigurations in getChoiceInputs
-        choices: Wheel.getChoiceInputs(gameToken.chainId, gameToken.affiliateHouseEdge).map(
-          (i) => ({
-            name: `${i.label} (${i.winChancePercent
-              .map(
-                (chance, index) =>
-                  `${i.formattedNetMultiplier?.[index]}x - ${chance.toFixed(2)}% to win ${index !== i.winChancePercent.length - 1 ? "|" : ""}`,
-              )
-              .join("\n")})`,
-            value: i,
-          }),
-        ),
+        choices: WeightedGame.getChoiceInputs(
+          gameToken.chainId,
+          gameToken.game,
+          gameToken.affiliateHouseEdge,
+        ).map((i) => ({
+          name: `${i.label} (${i.winChancePercent
+            .map(
+              (chance, index) =>
+                `${i.formattedNetMultiplier?.[index]}x - ${chance.toFixed(2)}% to win ${index !== i.winChancePercent.length - 1 ? "|" : ""}`,
+            )
+            .join("\n")})`,
+          value: i,
+        })),
       });
     }
   }
@@ -289,6 +296,20 @@ function _displayRolledBet(rolledBet: CasinoRolledBet) {
         commonMessage,
       ),
     );
+  }
+}
+
+export async function _refreshLeaderboardsWithBet(betId: string | bigint, chainId: CasinoChainId) {
+  console.log(chalk.blue(`Refreshing leaderboards with bet #${betId}...`));
+  const isSuccess = await wagmiBetSwirlClient.refreshLeaderboardsWithBets(
+    [String(betId)],
+    chainId,
+    LEADERBOARD_TYPE.CASINO,
+  );
+  if (isSuccess) {
+    console.log(chalk.green(`Leaderboards refreshed with bet #${betId} successfully!`));
+  } else {
+    console.log(chalk.red(`Failed to refresh leaderboards with bet #${betId}`));
   }
 }
 
@@ -425,7 +446,7 @@ export async function _placeBet(
   const commonParams = {
     betCount,
     betAmount,
-    token: casinoGameToken,
+    tokenAddress: casinoGameToken.address,
   };
   const callbacks = {
     onApprovePending: (_tx: Hash, _result: ApproveResult) => {
@@ -483,11 +504,15 @@ export async function _placeBet(
       casinoGameToken.chainId,
     );
   }
-  // Wheel
+  // Wheel & Plinko (Weighted game)
   else {
-    const weightedGameChoice = inputChoice as WheelChoiceInput;
-    placedBetData = await wagmiBetSwirlClient.playWheel(
-      { ...commonParams, weightedGameConfig: weightedGameChoice.config },
+    const weightedGameChoice = inputChoice as WeightedGameChoiceInput;
+    placedBetData = await wagmiBetSwirlClient.playWeightedGame(
+      {
+        ...commonParams,
+        weightedGameConfig: weightedGameChoice.config,
+        game: weightedGameChoice.game as WEIGHTED_CASINO_GAME_TYPE,
+      },
       undefined,
       callbacks,
       casinoGameToken.chainId,
@@ -651,11 +676,15 @@ export async function _placeFreebet(
       callbacks,
     );
   }
-  // Wheel
+  // Wheel & Plinko (Weighted game)
   else {
-    const weightedGameChoice = inputChoice as WheelChoiceInput;
-    placedBetData = await wagmiBetSwirlClient.playFreebetWheel(
-      { ...commonParams, weightedGameConfig: weightedGameChoice.config },
+    const weightedGameChoice = inputChoice as WeightedGameChoiceInput;
+    placedBetData = await wagmiBetSwirlClient.playFreebetWeightedGame(
+      {
+        ...commonParams,
+        weightedGameConfig: weightedGameChoice.config,
+        game: weightedGameChoice.game as WEIGHTED_CASINO_GAME_TYPE,
+      },
       undefined,
       callbacks,
     );