+++ packages/core/src/actions/index.ts
@@ -1,12 +1,16 @@
 /* Casino */
+
+export * from "../read/common/tokenMetadata";
+export * from "./api/leaderboard/leaderboards";
 export * from "./casino/cointoss";
 export * from "./casino/dice";
-export * from "./casino/roulette";
-export * from "./casino/keno";
 export * from "./casino/game";
-export * from "./casino/wheel";
+export * from "./casino/keno";
+export * from "./casino/plinko";
+export * from "./casino/roulette";
 export * from "./casino/weightedGame";
-
+export * from "./casino/wheel";
 /* Common */
 export * from "./common/approve";
-export * from "../read/common/tokenMetadata";
+/* Leaderboard */
+export * from "./leaderboard/leaderboard";