+++ packages/core/src/entities/casino/keno.ts
@@ -1,7 +1,5 @@
-import { type EncodeAbiParametersReturnType, encodeAbiParameters } from "viem";
-import { parseAbiParameters } from "viem";
-import { getFormattedNetMultiplier } from "../..";
-import { CASINO_GAME_TYPE, getNetMultiplier } from "../..";
+import { type EncodeAbiParametersReturnType, encodeAbiParameters, parseAbiParameters } from "viem";
+import { CASINO_GAME_TYPE, getFormattedNetMultiplier, getNetMultiplier } from "../..";
 import { BP_VALUE } from "../../constants";
 import type { KenoConfiguration } from "../../read/casino/keno";
 import { AbstractCasinoGame, type ChoiceInput } from "./game";
@@ -84,7 +82,7 @@ export class Keno extends AbstractCasinoGame<
     selectedBallsCount: number,
     matchedBallsCount: number,
   ): number {
-    return kenoConfig.mutliplierTable[selectedBallsCount]?.[matchedBallsCount] || 0;
+    return kenoConfig.multiplierTable[selectedBallsCount]?.[matchedBallsCount] || 0;
   }
 
   static getFormattedMultiplier(